beging moving s3 client intialization to apollo_gql
