+++ services/app-api/src/resolvers/configureResolvers.ts
@@ -41,13 +41,15 @@ import { contractRevisionResolver } from './contract/contractRevisionResolver'
 import { fetchContractResolver } from './contract/fetchContract'
 import { submitContract } from './contract/submitContract'
 import { rateRevisionResolver } from './rate/rateRevisionResolver'
+import type { S3ClientT } from '../s3'
 
 export function configureResolvers(
     store: Store,
     emailer: Emailer,
     emailParameterStore: EmailParameterStore,
     launchDarkly: LDService,
-    jwt: JWTLib
+    jwt: JWTLib,
+    s3Client: S3ClientT
 ): Resolvers {
     const resolvers: Resolvers = {
         Date: GraphQLDate,
@@ -143,7 +145,7 @@ export function configureResolvers(
         RateRevision: rateRevisionResolver,
         Contract: contractResolver(),
         ContractRevision: contractRevisionResolver(store),
-        GenericDocument: genericDocumentResolver(),
+        GenericDocument: genericDocumentResolver(s3Client),
     }
 
     return resolvers
