+++ services/app-api/src/resolvers/shared/genericDocumentResolver.ts
@@ -8,7 +8,7 @@ import {
 import type { S3BucketConfigType } from '../../s3/s3Amplify'
 import type { S3ClientT } from '../../s3'
 
-export function genericDocumentResolver(): Resolvers['GenericDocument'] {
+export function genericDocumentResolver(s3Client: S3ClientT): Resolvers['GenericDocument'] {
     return {
         downloadURL: async (parent) => {
             try {
@@ -20,42 +20,6 @@ export function genericDocumentResolver(): Resolvers['GenericDocument'] {
                         'S3 needs to be provided a valid key and bucket'
                     )
                     throw err
-                }
-                const s3LocalURL = process.env.REACT_APP_S3_LOCAL_URL
-                const s3DocumentsBucket =
-                    process.env.REACT_APP_S3_DOCUMENTS_BUCKET
-                const s3QABucket = process.env.REACT_APP_S3_QA_BUCKET
-                console.info(
-                    '============ s3DocumentsBucket ==============',
-                    s3DocumentsBucket
-                )
-                console.info(
-                    '============ s3QABucket ==============',
-                    s3QABucket
-                )
-                if (
-                    s3DocumentsBucket === undefined ||
-                    s3QABucket === undefined
-                ) {
-                    throw new Error(
-                        'To configure s3, you  must set REACT_APP_S3_DOCUMENTS_BUCKET and REACT_APP_S3_QA_BUCKET'
-                    )
-                }
-
-                let s3Client: S3ClientT
-                const S3_BUCKETS_CONFIG: S3BucketConfigType = {
-                    HEALTH_PLAN_DOCS: s3DocumentsBucket,
-                    QUESTION_ANSWER_DOCS: s3QABucket,
-                }
-
-                if (process.env.REACT_APP_AUTH_MODE !== 'LOCAL') {
-                    s3Client = newAmplifyS3Client(S3_BUCKETS_CONFIG)
-                } else if (s3LocalURL) {
-                    s3Client = newLocalS3Client(s3LocalURL, S3_BUCKETS_CONFIG)
-                } else {
-                    throw new Error(
-                        'You must set either REACT_APP_S3_REGION or REACT_APP_S3_LOCAL_URL depending on what environment you are in'
-                    )
                 }
 
                 const url = await s3Client.getURL(key, 'HEALTH_PLAN_DOCS')
